Add all initial website directories and files
Add directories 'images', 'recipes' (including all recipe sub-sites) and
'index.html' as the main website, as well as a modified "README.md"
file.

The formerly empty repository which was created on GitHub and cloned on
this local machine now includes all necessary directories and files for
a functioning website. The website currently contains three recipes and
one website for each recipe respectively.

Further changes should be committed as atomic commits so bugs and
versions can be easily traced.